Seperate apply function
