+++ package.json
@@ -25,6 +25,7 @@
     "elite-moodle-scraper": "^1.0.5",
     "express": "^4.16.3",
     "mongoose": "^5.2.13",
+    "node-schedule": "^1.3.0",
     "passport": "^0.4.0",
     "passport-google-auth": "^1.0.2"
   }
